requirements.txt
flask==3.1.2
supabase==2.7.4
python-dotenv==1.0.0
bcrypt==4.2.0
PyJWT==2.8.0
